+++ src/builders/requestBuilders/updates/updates.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-classes-per-file */
 export abstract class UpdateExpr {
-  abstract toQuery(): string;
+  abstract toQuery(position?: number): { query: string, values: Array<any> };
 }
 
 export abstract class UpdateCustomExpr<T> extends UpdateExpr {