+++ src/builders/requestBuilders/where/where.ts
@@ -1,3 +1,5 @@
+/* eslint-disable max-classes-per-file */
 export default abstract class Expr {
-  abstract toQuery(): string;
+  abstract toQuery(position?: number, tableCache?: {[tableName: string]: string})
+  : { query: string, values: Array<any> };
 }